poc/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
httpx==0.27.2
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
beautifulsoup4==4.12.3
Pillow==10.4.0
gradio_client>=1.3.0
